# openapi.yml
openapi: 3.1.0
info:
  title: CubeAPI
  description: OpenAPI contract for the CubeSandbox dashboard surface.
  license:
    name: ''
  version: 0.1.0
servers:
- url: /cubeapi/v1
  description: CubeAPI dashboard surface
paths:
...
  /tools:
    get:
      tags:
      - handlers::tools
      operationId: list_tools
      responses:
        '200':
          description: Tool list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ToolSummary'
        '500':
          description: Unexpected backend error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
    post:
      tags:
      - handlers::tools
      operationId: create_tool
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateToolRequest'
        required: true
      responses:
        '201':
          description: Tool created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ToolDetail'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '500':
          description: Unexpected backend error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /tools/{toolID}:
    get:
      tags:
      - handlers::tools
      operationId: get_tool
      parameters:
      - name: toolID
        in: path
        description: Tool identifier
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Tool detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ToolDetail'
        '404':
          description: Tool not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '500':
          description: Unexpected backend error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
    delete:
      tags:
      - handlers::tools
      operationId: delete_tool
      parameters:
      - name: toolID
        in: path
        description: Tool identifier
        required: true
        schema:
          type: string
      responses:
        '204':
          description: Tool deleted
        '404':
          description: Tool not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '500':
          description: Unexpected backend error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
    patch:
      tags:
      - handlers::tools
      operationId: update_tool
      parameters:
      - name: toolID
        in: path
        description: Tool identifier
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateToolRequest'
        required: true
      responses:
        '200':
          description: Tool updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ToolDetail'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: Tool not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '500':
          description: Unexpected backend error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
...
components:
  schemas:
    ApiError:
      type: object
      required:
      - code
      - message
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string
...
    CreateToolRequest:
      type: object
      properties:
        annotations:
          type:
          - object
          - 'null'
          additionalProperties:
            type: string
          propertyNames:
            type: string
        default_timeout:
          type:
          - integer
          - 'null'
          format: int32
        instance_type:
          type:
          - string
          - 'null'
        labels:
          type:
          - object
          - 'null'
          additionalProperties:
            type: string
          propertyNames:
            type: string
        name:
          type:
          - string
          - 'null'
        network_type:
          type:
          - string
          - 'null'
        runtime_handler:
          type:
          - string
          - 'null'
        storage_mounts:
          type: array
          items:
            $ref: '#/components/schemas/ToolStorageMount'
        template_id:
          type:
          - string
          - 'null'
        tool_id:
          type:
          - string
          - 'null'
...
    ToolDetail:
      type: object
      required:
      - tool_id
      - template_id
      properties:
        annotations:
          type:
          - object
          - 'null'
          additionalProperties:
            type: string
          propertyNames:
            type: string
        default_timeout:
          type:
          - integer
          - 'null'
          format: int32
        instance_type:
          type:
          - string
          - 'null'
        labels:
          type:
          - object
          - 'null'
          additionalProperties:
            type: string
          propertyNames:
            type: string
        name:
          type:
          - string
          - 'null'
        network_type:
          type:
          - string
          - 'null'
        runtime_handler:
          type:
          - string
          - 'null'
        storage_mounts:
          type: array
          items:
            $ref: '#/components/schemas/ToolStorageMount'
        template_id:
          type: string
        tool_id:
          type: string
    ToolHostDirStorageSource:
      type: object
      required:
      - host_path
      properties:
        host_path:
          type: string
    ToolStorageMount:
      type: object
      required:
      - name
      - mount_path
      - storage_source
      properties:
        mount_path:
          type: string
        name:
          type: string
        read_only:
          type:
          - boolean
          - 'null'
        storage_source:
          $ref: '#/components/schemas/ToolStorageSource'
        sub_path:
          type:
          - string
          - 'null'
    ToolStorageSource:
      type: object
      properties:
        host_dir:
          oneOf:
          - type: 'null'
          - $ref: '#/components/schemas/ToolHostDirStorageSource'
    ToolSummary:
      type: object
      required:
      - tool_id
      - template_id
      properties:
        instance_type:
          type:
          - string
          - 'null'
        name:
          type:
          - string
          - 'null'
        network_type:
          type:
          - string
          - 'null'
        template_id:
          type: string
        tool_id:
          type: string
  securitySchemes:
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...